@@ -15,7 +15,6 @@
     "@types/node": "^12.12.54",
     "@types/radium": "^0.24.3",
     "@types/react": "^16.9.46",
-    "@types/react-dom": "^16.9.8",
     "@types/react-helmet": "^6.1.0",
     "@types/react-router": "^5.1.8",
     "@types/react-router-dom": "^5.1.6",
@@ -43,8 +42,8 @@
     "primereact": "^7.2.1",
     "query-string": "^6.13.1",
     "quill": "^1.3.7",
-    "react": "^17.0.1",
-    "react-dom": "^17.0.1",
+    "react": "^18.0.0",
+    "react-dom": "^18.0.0",
     "react-helmet": "^6.1.0",
     "react-leaflet": "^3.1.0",
     "react-rnd": "^10.3.5",
@@ -87,12 +86,13 @@
     "not op_mini all"
   ],
   "devDependencies": {
-    "react-scripts": "^4.0.0",
     "@types/leaflet": "^1.5.19",
     "@types/quill": "^1.3.10",
+    "@types/react-dom": "^18.0.0",
     "@types/resize-observer-browser": "^0.1.4",
     "babel-core": "^6.26.3",
     "babel-runtime": "^6.26.0",
+    "react-scripts": "^4.0.0",
     "webpack-cli": "^4.2.0"
   },
   "files": [
