@@ -1,21 +1,13 @@
-/** React imports */
 import React from 'react';
-import ReactDOM from 'react-dom';
-
-/** scss */
 import './index.scss';
-
-/** Other imports */
 import * as serviceWorker from './serviceWorker';
 import MiddleMan from './MiddleMan';
 
-/** Rendering the application */
-export default ReactDOM.render(
-  // <React.StrictMode>
-  <MiddleMan/>,
-  // </React.StrictMode>,
-  document.getElementById('root')
-)
+import { createRoot } from 'react-dom/client';
+
+const container = document.getElementById('root');
+const root = createRoot(container!); // createRoot(container!) if you use TypeScript
+root.render(<MiddleMan/>);
 
 // If you want your app to work offline and load faster, you can change
 // unregister() to register() below. Note this comes with some pitfalls.
